Methods z_mid,all 't', beta,mr.gamma,hc written.

+++ RainCalcs/RainVariables.cs
@@ -8,59 +8,59 @@ namespace CoreWpfApp.RainCalcs
     {
         //Variables initialization
 
-        internal double F; //ok+
-        internal double F_road; //ok+
-        internal double F_green;  //ok+
-        internal double F_roof; //ok+
-        internal double F_gravel; //ok+
-        internal double n; //Pril2 table +
-        internal double A;  //method+
-        internal double Psy_mid; //method+
-        internal double Z_mid; //method+
+        internal decimal F; //ok+
+        internal decimal F_road; //ok+
+        internal decimal F_green;  //ok+
+        internal decimal F_roof; //ok+
+        internal decimal F_gravel; //ok+
+        internal decimal n; //Pril2 table +
+        internal decimal A;  //method+
+        internal decimal Psy_mid; //method+
+        internal decimal Z_mid; //method+
         public int N_sections;
-        internal double t_con; //ok+
-        internal double t_can; //method+
-        internal double t_p; //method+
-        internal double l_lotok; //ok+
-        internal double v_lotok; //ok+
-        internal double l_p; //ok+
-        internal double v_p; //ok+
-        internal double Q_rRain; //method+
-        internal double Q_rIce; //method+
-        internal double beta; //table+
-        internal double Q_cal; //method+
+        internal decimal t_con; //ok+
+        internal decimal t_can; //method+
+        internal decimal t_p; //method+
+        internal decimal l_lotok; //ok+
+        internal decimal v_lotok; //ok+
+        internal decimal l_p; //ok+
+        internal decimal v_p; //ok+
+        internal decimal Q_rRain; //method+
+        internal decimal Q_rIce; //method+
+        internal decimal beta; //table+
+        internal decimal Q_cal; //method+
         internal int Q_20; //table+
-        internal double P; //table+
-        internal double m_r; //table
-        internal double gamma; //table
-        internal double K; //table
-        internal double Z_i; //table
-        internal double Psy_i; //table
+        internal decimal P; //table+
+        internal decimal m_r; //table
+        internal decimal gamma; //table
+        internal decimal K; //table
+        internal decimal Z_i; //table
+        internal decimal Psy_i; //table
         internal int h_c; //table
-        public double t_c() //
+        public decimal t_c() //
         {
              return t_p;
         }
-        public double K_y() // -> METHOD 21072020+
+        public decimal K_y() // -> METHOD 21072020+
         
         {
-            double _F_y = F_y();
+            decimal _F_y = F_y();
             return 1 - (_F_y / F);
         }
-        public double F_y() // -> METHOD 21072020+
+        public decimal F_y() // -> METHOD 21072020+
         
         {
             return F_road;
         }
 
-        public double F_roadRoof() //-> METHOD 21072020+
+        public decimal F_roadRoof() //-> METHOD 21072020+
         
         {
             return F_road + F_roof;
             
         }
 
-        public double t_r() //-> METHOD 21072020+
+        public decimal t_r() //-> METHOD 21072020+
         {
              return t_con + t_can + t_p;
         }
